Karapace: Initial commit to get things started
Karapace is an aiohttp based daemon written in Python 3.6
that allows you initially to store and retrieve schemas.

Also schema compatibility testing support for the modes
BACKWARD, FULL and FORWARD compatibility exist, though
could surely be more comprehensive.

There's a fairly extensive set of unittests for the project
but there are sure to be issues that remain so we'll need
even more of them in the future.

The initial release is roughly ~2.5k LOC in all.