compatibility: extend supported types
Compatibility checks are extended to support Union, Array, Map and Fixed
types both directly and nested. Primitive and Enum schemas can now also
be used directly and not only in a nested record structure. The simple
presentation for types in maps and arrays is also made to look at
compatibility promotions correctly. In addition, the test suite is
extended with more schema checks to increase coverage.