bugfix: fetch all schemas from a large topic
Because poll works in batches, topics with more than `max_poll_records`
(default of 500) have to be polled multiple times, otherwise only the
first `max_poll_records` are saved by the backup tool.

This assumes there are two timeouts, one for the request and another for
the socket read. A request timeout would occure when the server does not
have additional data to send the client after the timeout, so an empty
response is returned by the server. A socket read timeout happenes when
the client is waiting for the server but a response never arrives.

The above is necessary because it allows to use an empty response as an
indicator of "no more data".