jsonschema: clarified error message
The previous error message was correct, but it lacked enough information
for the user to understand why the addition of a new property is a
backwards incompatible change. This tries to fix the issue by explicitly
mentioning the lack of validation for those properties, and point the
user to additionalProperties/patternProperties.