Fix cluster_metadata caching and locking
In case cluster_metadata was fetched first for a limited set of
topics, the method internally cached partial metadata, which was then
returned regardless if full metadata was requested.

Also use time.monotonic() for cache expiry so that difference is
always valid.

Adjust test configuration for ability to test caching issues.