bugfix: hotloop on get_offset_from_queue
Under some circumstances it may happen that an unexpected offset is
added to the Queue. When that happens, there is no thread to consume the
offset, so the offset is `get`ed and immediately `put`ed back. The
result is that `get_offset_from_queue` becomes a hot loop and spams the
logging system and slows Karapace as a whole.

Additionally, when dozens of offsets are added to the Queue, each
iteration of the loop will get the element from the top of the queue,
check it, and if it is not the target offset put it back at the end of
the queue. The result is that the method degrades to a very slow linear
search.

This commit does not fix the bug that introduces an unexpected offset to
the Queue, but it does introduced a few fixes:

- Fixes the hotloop, with this change a single Event object is used for
  each offset.
- Adds a timeout for the waiting. Not seeing the expected offset may be
  another issue (yet to be confirmed).
- Removes the linear search by an Event object.