tests: fix integration tests flakiness
This contains a variety of chances to remove flakiness of the
integration tests. There were two main culprits for flakiness:

1: Using the same topic for the schemas:

- Each test has its own registry, when running tests in parallel, to
  prevent that these servers join the same group, different group ids
  are used.
- However, the registries were all using the _schemas topic to store the
  schemas, which meant there where multiple nodes thinking they are the
  leader writing to the same topic, causing flakiness in the tests
    - This corrupted the schema IDs, meaning a different test could
      write a totally different schema with the same id via another
      leader.

2: The same port range was used for all runners, which meant port reuse
   was possible

Besides fixing the above, this also changes how registries are executed
in the test suite. With this change the servers are started using
subprocesses. This has the disavantage that cleanup is harder, but it
does test the initialization and server codepath, which previously was
skipped because of the aiohttp client.