website: Add Sphinx website with a base for docs
This converts the website to a Sphin site that
contains a landing page and an initial setup
for docs included in the same site that works
as a base for further documentation development.

The styling for the docs content is very
minimal with the basic setup that comes from the
Furo theme and a few additional color variables,
logos etc. that are meant to be further customized.