fix: Schema registry message key field format correction
Fixes Karapace issue #347

Confluent Schema Registry message key uses field order of
`keytype, subject, version, magic` and Karapace has used order
of `subject, version, magic, keytype`.
Kafka won't compact the topic correctly as the key field byte contents are different.

On Karapace startup the schema message topic is consumed fully. The the key correction
is started in background thread and each key format checked. If format is not correct
a tombstone message and corrected message is sent to the topic. Records that have valid
key format the same actions are done, tombstone and the record with original key and value
are sent. This is to retain order of actions.

In the end a special record is sent to schemas topic to mark that key correction has been
run. This serves also as a marker for follower registries not to start correction on
leader change.

Kafka will ultimately compact the topic and remove the incorrectly formatted entry and
the tombstone.