Acquire lock once per batch
This is an optimization.

In the case of a batch of e.g. 500 records, it makes sense to acquire and release self.id_lock only once, because locking is relatively expensive operation.

This will naturally mean that other lock user cannot acquire the lock while the batch is being processed. This is not an issue, it is actually a good thing. The others will get fresh data, since the locking in KafkaSchemaReader typically means there's going to be changes in the schemas.