Primary node selection prefers ready node
The primary node selection prefers node that is ready, has replayed
the schemas topic. The selection uses the strategy, i.e. lowest ready
or highest ready. Selected node will stay as primary until a new
group selection is run and new (or the same) node is selected as primary
based on the strategy.

Fallback will be to select primary based on the strategy from the
full set of nodes.