fix: Karapace instance primary eligiblity added to group protocol
Group coordinator leader is selected from the whole set of members
of the group. This can also be member that is not primary eligible.
This change corrects the group protocol data and Karapace primary
selection to use the actual instance configured primary eligibility.
The code before used the primary eligibility of the selected group
leader which can be false.