Fix `allocate_port()` on Mac
The `PortRangeInclusive.allocate_port()` function uses `psutil` under the hood
to find out if a port is allocatable. This leads to an `AccessDenied` error on
Mac. Effectively making it impossible to run any of the integration tests.

This patch changes the implementation to not rely on `psutil` at all, it also
removes it from the dependencies since this is the only usage. Instead, we
simply try to bind to a port and see if it is free, while keeping track of the
ports we already predestined for binding by one of our own components.

This does not fix any possible race conditions that might arise from trying to
find a free port on a system where other processes are running that might bind
to ports. There is no way to achieve this since any free port can be bound by
any process. The only way to achieve this would be if the software in question
itself has direct support for finding a suitable port. It would be possible to
add this functionality to Karapace (e.g., if port is set to `-1` automatically
find a suitable one). However, I did not go in that direction, as it would
require extensive changes to various parts. The patch is only meant to allow
integration test execution on Mac, without any visible API changes.

A detail I think worth mentioning that might not be clear from the patch itself
is that it removes the `maybe_available` collection in favor of a `bound`
collection. The `maybe_available` collection is created from the given `start`
and `end` parameters, that range might be big (up to 64,511 integers to be
precise) and consume an unreasonable amount of memory. The `bound` approach
does the opposite of the `maybe_available` and keeps track of those ports that
were identified as being free, hence, the collection is small throughout the
entire execution time.

Last, but not least, the `port_is_listening()` was inline as it had only a
single usage and the `ipv6` flag was removed as it was unused. IPv6 support
would require a more global approach for enabling if needed.