Fix `make copyright`
The `make copyright` goal always lists all Python files that have a copyright
notice, and since we have more than one file with a copyright notice it always
exits with success. This does not seem right, given that there is a GHA job that
is supposed to check that all files (except `__init__.py`) have a copyright
notice. This changes the code used in the goal to a single git invocation that
does exactly that. The `!` in front inverts the exit code. Hence, if a Python
file other than `__init__.py` is found without a copyright notice it exists with
a non-zero exit code and lists the paths that miss the copyright notice relative
to the repository root.