Add Backup File Overwrite Protection
The schema backup tool is currently overwriting the output file without asking
to do so. It is also writing directly to the output file, and leaves the file
intact even if the backup actually failed. This patch adds guards that fail if
the output file already exists, but also an `--overwrite` flag. It also
introduces a temporary file that is persisted only if the backup actually
succeeded, not leaving behind any trace of the work if not.

The output needs to be a file to be overwritten, even with the `--overwrite`
flag present. Any other path type is not going to be touched. Our code verifies
this, and Python's `replace` function does as well.

Other smaller things worth mentioning:

- The `--location` accepts `-` as synonym for `stdout` as is conventional, the
  current empty string continues to work as well and stays the default. But, `-`
  is not going to create a file called `-` anymore.
- Renamed `Writer` to `_writer` and by doing so made it private. This could be
  considered a breaking change since the name of the symbol changed, and anyone
  relying on it fails now. However, the `Writer` was never meant to be a public
  symbol, and we do not have any policy regarding such changes.
- Fixed the return type hint from `Iterable[TextIO]` to `TextIO` since it is a
  context manager, and not something that can be iterated.