Add Backup Partition Count Check
This adds a check that verifies that the source and target topic for a backup
indeed have a single partition, instead of simply assuming that it does. This is
not a breaking change, since it makes the implicit explicit, and users who are
going to receive the error are going to be protected from a lot of hardship.

This also removes the consumer and producer clients from the class by moving
them into dedicated top-level functions. Those top-level functions are context
managers that close the clients automatically in a reliable way. This removal
could be considered a breaking change, since we are removing public functions
from a class. However, these functions were never meant to be public, especially
not because they were directly meddling with internal state.

The new top-level functions **always** check the partition count, and client
creation without a partition count check is not possible. This is on purpose,
to ensure safety. Last, but not least, there are tests that verify all of the
claims.