Fix Backup Topic Creation Retries and Interrupts
The retry logic in the backup code for admin client and topic creation is flawed
as it compares seconds to milliseconds. Instead of aborting after the desired
60 seconds it aborts after 60,000 seconds. The retry implementation itself has
several other issues, most notable the fact that it swallows `BaseException`,
and thus cannot be interrupted by the user when it enters the retry loop.

Instead of trying to come up with an optimized retry logic or a generic solution
that can be used across the board, I decided to add a well tested library that
already provides this logic.

This updates only the backup code. Other code might or might not have similar
flaws. For instance, after this change there are still 11
`pylint: disable=bare-except` left in the `karapace` source directory.