Fix Python Executable Search in Tests
This removes the hardcoded `python` executable name in tests with a more dynamic
approach that resolves to the Python executable used by the tests. Not all
systems have Python 3 available at `python` by default, and there is no
guarantee that the `PATH` in a `venv` is modified to include the `venv/bin`
directory in the `PATH`. Python offers `sys.executable` to find the current
executable, if that fails we fall back to the same environment variable we use
in our make file, and if that fails we use `python3` (seems to be available on
most systems).