feat: Add Backup Format 3 Protobuf Schema
This adds the Protobuf schema for the version 3 backup format. Protobuf is going
to enable us to statically verify backward and forward compatibility of the
schema, and thus allow us to add fields without introducing a new version in the
future. Protobuf uses a universal C extension for their dynamic languages,
resulting in a massive speed-up. This is available starting with Protobuf 20, we
thus will need to upgrade Protobuf to actually make use of this.