feat: Add `shellcheck` and `shfmt` Pre-Commit Hooks
This adds `shellcheck` and `shfmt` pre-commit hooks. I fixed the most pressing
issues that were found, but did not go through the existing scripts otherwise.
We are going to need more shell scripts to automate other things, this is why it
is a good idea to add these tools to our verification chain.

~This also changes the GHA lint workflow to use the official pre-commit action,
which also makes use of caching, and thus should be faster.~ The pre-commit
action uses the `--color=always` option, and that option results in the action
hanging forever if paired with local hooks, which we have.