build: Add Basic Build System
This adds a basic build system that allows developers to get started with
Karapace without the need of figuring out what is required to work with the
project on their own. The README documents how to use it, as well as all quirks
that are known to me at this time.

Closes #553