Introduce narrower JSON typing
This introduces narrower type aliases to describe JSON-compatible dicts.
Because we now have more accurate typing, in some places this reveals
places where we are not properly dealing with all possible types that
deserialized JSON can have. In order for this PR to not do too much in
one go, I opted to introduce `TypedDict`s along with a mechanism to cast
directly invocations of `json_decode()`. This reflects the same
assumptions we make in runtime onto to our static typing.

We should consider introducing tooling to allow properly deserializing
into rich Python types, such as Pydantic or Dacite.