feature: Implement V3 backup format
A new backup format is introduced as a set of new backup backends. The
new format uses Avro for serialization and parsing. To achieve a chain
of type safety through on-disk serialization, Python runtime, and static
type checking, we use Python dataclasses with type annotations and infer
Avro representations from those.

Avro schemas are checked-in into version control, and checked against
previous versions using a CI job, to make sure we cannot introduce
breaking changes. The CI job also checks that the dataclasses are in
sync with checked-in schema.

The new format stores a lot more data than the previous versions, both
on the record-level, as well as on the topic/partition level. It uses an
approach to checksumming that allows tuning the level of precision in
determining where in a file a mismatch occured vs disk consumption.
Checksums are verified both after a full partition is written, as well
as for each checksum checkpoint found in a file.

A change in the new format is also that it is written to multiple files
instead of a single file. Because of this it doesn't support writing to
stdout. Similarly to current versions, it writes files to a temporary
directory in the target directory, then moves the temporary directory in
place once successful.