feature: Implement backup metadata inspect command
Implements a command for dumping metadata of a backup file:
`karapace_schema_backup inspect --location=<backup location>`.

Refactors the internal API into separate functions instead of the shared
`SchemaBackup` class. While this was done for the purpose of fitting the
inspect command, which does not make sense to require a config file for,
the existing methods did not gain much from having a shared class.