fix: Reduce mutability in schema references
First off, this fixes the issue that `Reference` objects were mutated
while acting as immutable by implementing `__hash__()`. To do that, it
breaks up the `Reference` class into two, so that instead of it having
to overloaded meanings, the unresolved state is instead represented as
its own entity, `LatestVersionReference`. This is a typical "make
illegal state unrepresentable" kind of refactoring.

Related to this, a few `list` and `dict`s are replaced with their
immutable abstract counterparts, `Sequence` and `Mapping`. This greatly
helps in reasoning about where mutation can happen, and stops me from
constantly wondering "will this list be altered when I pass this down
this chain of methods".