chore: Refactor various protobuf modules
The main goal is to increase type coverage.

The `Location` class is turned into a dataclass to simplify things and
gain immutability. It's `.get()` method needs an overload to type
accurately, even after simplifying its signature by raising instead of
returning `None` when incorrect amount of arguments are given. This was
addressed as a side-quest in adding types to `ProtobufSchema`, because
it has an attribute that is assigned to from the return value of this
method.

The `ProtoFileElement.__eq__()` is altered so that it doesn't assume the
compared object is of the same type as itself, and its argument is
updated to follow convention of supporting equality check with any other
type of object.

`ProtobufSchema.dirty` is removed because it was unused.

`ProtobufSchema.references` along with its init argument are removed
because they aren't used. This triggers a chain of removing unused
arguments in a few functions.