feature: Allow passing backup directory to backup commands
With this change, we allow passing the same path of the directory
created when producing a backup, into restoration and inspection
commands. Together with the fix in 5becdb4b, this makes V3 more similar
to V2, in the respect of handling given paths.

Also improves test coverage of some of the helper functions in the
internal API.