Resolve entities recursively when receiving the schema
The previous implementation did not correctly resolve entities within a schema that contained references. This led the REST Proxy responds with a 422 error even when the message was correct.
To address this issue, the schema resolver has been modified to ensure that schemas with references are recursively resolved and that the references are passed to the parser correctly.