Replace admin client with confluent-kafka one
This change replaces all Kafka admin client usages with a new
implementation, based on confluent-kafka-python's `AdminClient`, aiming
to keep the same interface as much as possible for the time being.

The goal is not to completely remove kafka-python as a dependency,
rather take a step towards that. The kafka-python library is still widely
used in the codebase, eg. for error handling (even in the new
`KafkaAdminClient`).

The new `KafkaAdminClient` takes the place of all previous admin clients
used, thus not specific to the REST proxy anymore. Hence it's move to a
new module as `karapace.kafka_admin.KafkaAdminClient`, which also
consolidates all references to the `confluent_kafka` lib. The test suite
has also been moved and expanded.

Wherever possible the same interface was kept, only adding new methods
to simplify client code (to a minimal extent). This is most notable in
tests and the backups functionality. The new admin client aims to hide
away confluent-kafka-python's Future-based async client approach by
providing wrapper methods that resolve these futures.

Typing (issues) from `confluent_kafka` are ignored for now, no typing
stubs have been added on purpose.

Resources:
* confluent-kafka-python documentation: https://docs.confluent.io/platform/current/clients/confluent-kafka-python/html/index.html#
* librdkafka configuration documentation: https://github.com/confluentinc/librdkafka/blob/master/CONFIGURATION.md