protobuf: improve protoc compatibility
Fix a bug in _deserialize_field that causes deserialization fail in certain cases, e.g. in the case there's a nested field that is a message or an enum. The issue is that the type is optional if type_name is specified. The code falsely expects type to always be present.

Add test test_protoc_serialized_schema_deserialize to test deserialization of protoc-serialized schemas. Test a few cases that are problematic, e.g. one with an enum and another with a nested message.