refactor: optimize the authorization check for get schema by id
Load the subjects associated with the schema id from in-memory database.
This list can be used as source for subjects to match user ACLs.

The ACL authorizator is added which can operate only on the user
dictionary and permission list. HTTP authorizer inherits this class
and add the file loading and HTTP authorization handling on top.
Unit tests for ACL authorizer are added.