Handle message errors in schema reader
Previously the errors that can surface when polling for messages have
not been handled in the schema reader. This commit fixes that and
specifically handles and error-logs group and topic authorization
exceptions. This mirrors the similar situation in the REST proxy's
consumer manager.

Other errors (`KafkaError` and subclasses) raised here will be handled
and logged as unexpected in the main message handler loop.

Prompted by https://github.com/Aiven-Open/karapace/issues/812.