chore: Refine confluent kafka typing
Current type hints allowed passing `None` for some parameters of this
API whereas the documentation actually specifies that's not the case.
See c8f3923886fcae84daa0c0840abb9297d7cea93d which fixes a previous
defect. Docstring of this API specifies that `message` and `offsets` are
mutually exclusive, but that it's also valid to omit both. By
strictening the type hints to only allow exactly this, we avoid
reintroducing the same or similar defects in the future.