schema-reader: Shutdown service if corrupt entries in `_schemas` topic
Previously, when we encounter errors within the `_schemas` topic, we would continue the
message loading and skip the problematic schema.
This is not ideal as it might leave the application with corrupt schema data and the side-effects could be grave.
What we do now is to kill the service, log the errors and allow a graceful shutdown.
We will follow this work by adding metrics for such cases.