rest-proxy: allow reading messages with avro-value and string-key
It is not uncommon that users produce messages with avro-value and string-key.
Since Karapace REST only supports one format type for deserializing both
the message key and the message value, I suggest to fallback to binary
deserializer for the key only when it fails to be read using avro format.