rest-proxy: allow reading messages with avro-value and string-key
It is not uncommon that Kafka users produce messages with String-key
and Avro-value. In this case Karapace REST typically fails to read
the message since it only supports one format for deserializing both
the key and the value.
For more convenience, I suggest to fallback to `binary` deserializer
for the key (only) when reading it using `avro` fails.