backups: Handle transiently incorrect partition count
When using the backups library to restore topics on KRaft, we are
observing the partition count check failing. Subsequent manual checks of
the partition count shows the correct count. This commit adds a fix to
make the restoration process anticipate that partition count metadata
will be initially incorrect but eventually converge on the correct
value.